[Bug] Player position not being properly loaded when coming from main menu.
This was caused by the recent change to support levels on the main menu. When going from the main menu to gameplay now, we go through the level_trans states. However, the logic to load the player position assumed that we went from main_menu directly to gameplay. We now cache the previous "core" state so that we coming out of level_trans_exit we can load the player position there (in the case that they are coming from main menu).